+++ packages/odata-uri-builder/src/index.ts
@@ -1,2 +1,8 @@
-export { ODataUriBuilder, ODataUriBuilderV2, ODataUriBuilderV4 } from "./internal";
+export * from "./ODataUriBuilderModel";
+
+export { ODataUriBuilderV2 } from "./v2/ODataUriBuilderV2";
+export { ODataUriBuilderV4 } from "./v4/ODataUriBuilderV4";
+export { ExpandingODataUriBuilderV2 } from "./v2/ExpandingODataUriBuilderV2";
+export { ExpandingODataUriBuilderV4 } from "./v4/ExpandingODataUriBuilderV4";
+
 export { QFilterExpression } from "@odata2ts/odata-query-objects";
