+++ packages/odata-uri-builder/src/internal.ts
@@ -1,6 +1,7 @@
 export * from "./ODataModel";
-export * from "./ODataUriBuilderBase";
-export * from "./ODataUriBuilder";
+export * from "./ODataUriBuilderModel";
+export { ODataUriBuilder } from "./ODataUriBuilder";
 export { ODataUriBuilderV2 } from "./v2/ODataUriBuilderV2";
 export { ODataUriBuilderV4 } from "./v4/ODataUriBuilderV4";
+export { ExpandingODataUriBuilderV2 } from "./v2/ExpandingODataUriBuilderV2";
 export { ExpandingODataUriBuilderV4 } from "./v4/ExpandingODataUriBuilderV4";
