+++ packages/odata-uri-builder/src/v4/ExpandingODataUriBuilderV4.ts
@@ -1,6 +1,6 @@
 import { QFilterExpression, QOrderByExpression, QueryObject } from "@odata2ts/odata-query-objects";
-import { EntityExtractor, ExpandingODataUriBuilderV4Model, ExpandType } from "../ODataUriBuilder";
-import { ODataUriBuilderBase, ODataUriBuilderConfig } from "../ODataUriBuilderBase";
+import { EntityExtractor, ExpandingODataUriBuilderV4Model, ExpandType } from "../ODataUriBuilderModel";
+import { ODataUriBuilder } from "../ODataUriBuilder";
 
 /**
  * Builder for expanded entities or entity collections.
@@ -11,10 +11,10 @@ export class ExpandingODataUriBuilderV4<Q extends QueryObject> implements Expand
     return new ExpandingODataUriBuilderV4<Q>(property, qEntity);
   }
 
-  private builder: ODataUriBuilderBase<Q>;
+  private builder: ODataUriBuilder<Q>;
 
   public constructor(property: string, qEntity: Q) {
-    this.builder = new ODataUriBuilderBase(property, qEntity, { expandingBuilder: true });
+    this.builder = new ODataUriBuilder(property, qEntity, { expandingBuilder: true });
   }
 
   public select(...props: Array<keyof Q | null | undefined>) {
